devtools: add tracepoint check in checkpatch
This patch adds a check in checkpatch tool, to check if a
tracepoint is present in any new function added in cryptodev
library.

It uses the existing build_map_changes() function and version.map
file to create a list of newly added functions. The definition of
these functions are checked if they contain tracepoint.
The checkpatch return error if the tracepoint is not present.

For functions for which trace is not needed, they can be
appended to devtools/trace-skiplist.txt file. The above tracepoint
check will be skipped for them.

Signed-off-by: Ankur Dwivedi <adwivedi@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>