devtools: add tracepoint check in checkpatch
This patch adds a validation in checkpatch tool, to check if a
tracepoint is present in any new function added in cryptodev
library.

It uses the existing build_map_changes function to create a map
of functions. In the map, the newly added functions are identified
and their definition are checked for the presence of tracepoint.
The checkpatch return error if the tracepoint is not present.

For functions for which trace is not needed, they can be added to
devtools/trace-skiplist.txt file. The above tracepoint check will
be skipped for them.

Signed-off-by: Ankur Dwivedi <adwivedi@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>