net/iavf: revert fix VLAN insertion
When the kernel driver tells to use the L2TAG2 field for VLAN insertion,
the context descriptor needs to be used. There is an issue on the vector Tx
path, because it does not support the context descriptor.

The previous commit forces to select normal path to avoid the above issue,
but it results in a performance loss of around 40%. So it needs to be
reverted and the original issue needed to be fixed by rework.

To reverts
commit 0d58caa7d6d1 ("net/iavf: fix VLAN insertion")

Fixes: 0d58caa7d6d1 ("net/iavf: fix VLAN insertion")

Signed-off-by: Yiding Zhou <yidingx.zhou@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>