event/dlb2: fix port COS range allocation
This commit fixes allocation of port COS
when application requested port COS exceeds
(e.g. beyond 0-15) the number of LDB ports for
the domain.

We limit application specified ports from a
COS to the max ports allocated for the COS
so that the rest of the of the ports can be
allocated from default(best) COS.

Fixes: bec8901bfe9f ("event/dlb2: support ldb port specific COS")
Cc: stable@dpdk.org

Signed-off-by: Abdullah Sevincer <abdullah.sevincer@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>