eventdev/eth_tx: fix queue delete logic
To delete all the queues of a ethernet device associated with
adapter instance the queue_id can be passed as -1 to the queue
delete API.

When a subset of queues of a ethernet device are associated,
the queue delete logic is exiting without deleting the queues
in some cases (higher numbered associated queues) for above
scenario as the queue delete logic is not checking all the
queue association status.

This patch fixes this issue by checking the queue association
status of all the queues of the ethernet device.

Fixes: 741b499e642 ("eventdev/eth_tx: fix queue delete logic")
Cc: stable@dpdk.org

Signed-off-by: Naga Harish K S V <s.v.naga.harish.k@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>