app/testpmd: fix testpmd receive jumbo frame packets
For NIC I40E_10G-10G_BASE_T_X722, when testpmd is configured with
link speed, it cannot receive jumbo frame packets.

Because it has changed the link status of the ports if it was
configured with link speed.

When exiting testpmd that it automatically stops packet forwarding
and stops all the ports. But it doesn't update the link status of the
ports. If stop the ports first that it will update the link status.

This patch fix the error that testpmd will update the link status of
the ports when it exits.

Fixes: d3a274ce9dee ("app/testpmd: handle SIGINT and SIGTERM")
Fixes: 284c908cc588 ("app/testpmd: request device removal interrupt")
Cc: stable@dpdk.org

Signed-off-by: Jie Wang <jie1x.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>