examples/vdpa: support running in nested virtualization environment
When we run dpdk vdpa in the nested virtual machine vm-L1 and ping
test in vm-L2, the ping is NG. The reason for troubleshooting is
that the virtio net in vm-L2 sends control information to the vring,
and the qemu back-end device in vm-L1 cannot obtain correct data
from the vring. This problem is related to the opening of the vIOMMU.

This patch add option '--iommu-support' to use guest vIOMMU to
protect vhost, then the ping test in vm-L2 is OK.
This option is required in a nested virtualization environment.

Signed-off-by: Hao Chen <chenh@yusur.tech>
Signed-off-by: 0-day Robot <robot@bytheb.org>